whitebox_attack.py

+++ whitebox_attack.py
@@ -343,7 +343,7 @@ def main(args):
                     adv_text = tokenizer.decode(adv_ids)
                     x = tokenizer(adv_text, max_length=256, truncation=True, return_tensors='pt')
                     #token_errors.append(wer(adv_ids, x['input_ids'][0]))
-                adv_logit = model(input_ids=x['input_ids'].cuda(), attention_mask=x['attention_mask'].cuda())[0].data.cpu()
+                adv_logit = model(input_ids=torch.LongTensor(x['input_ids']).cuda(), attention_mask=x['attention_mask'].cuda())[0].data.cpu()
                 #print(adv_logit.size(),type(label))
                 #choice_list.append([adv_logit[0][label],adv_text])
                 if adv_logit[0][label] > ma[0][label]:
